3dpot-platform/services/api-gateway/requirements.txt
# 3dPot API Gateway Dependencies
# Criado em: 2025-11-12 22:42:43
# Autor: MiniMax Agent

# FastAPI e ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0  # PostgreSQL async
alembic==1.13.0

# Cache
redis[hiredis]==5.0.1
cachetools==5.3.2

# Object Storage
minio==7.2.0

# Message Queue
pika==1.3.2  # RabbitMQ

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# MQTT
paho-mqtt==1.6.1

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
orjson==3.9.10

# Logging
structlog==23.2.0
rich==13.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0